reorganize dielectric tensor
